requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asn1crypto==1.5.1
    # via -r requirements.in
astroid==2.15.6
    # via pylint
beautifulsoup4==4.12.2
    # via bs4
black==23.7.0
    # via -r requirements.in
blinker==1.6.2
    # via flask
bs4==0.0.1
    # via -r requirements.in
certifi==2024.7.4
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   black
    #   flask
cryptography==43.0.1
    # via
    #   -r requirements.in
    #   jwcrypto
dill==0.3.7
    # via pylint
flask==2.3.3
    # via -r requirements.in
idna==3.4
    # via requests
isort==5.12.0
    # via pylint
itsdangerous==2.1.2
    # via flask
jinja2==3.1.4
    # via flask
jwcrypto==1.5.6
    # via -r requirements.in
lazy-object-proxy==1.9.0
    # via astroid
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.2
    # via black
platformdirs==3.10.0
    # via
    #   black
    #   pylint
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via -r requirements.in
pykcs11==1.5.12
    # via -r requirements.in
pylint==2.17.5
    # via -r requirements.in
pyqt6==6.5.2
    # via
    #   -r requirements.in
    #   pyqt6-webengine
pyqt6-qt6==6.5.2
    # via pyqt6
pyqt6-sip==13.5.2
    # via
    #   pyqt6
    #   pyqt6-webengine
pyqt6-webengine==6.5.0
    # via -r requirements.in
pyqt6-webengine-qt6==6.5.2
    # via pyqt6-webengine
requests==2.32.2
    # via -r requirements.in
soupsieve==2.5
    # via beautifulsoup4
tomlkit==0.12.1
    # via pylint
typing-extensions==4.10.0
    # via jwcrypto
urllib3==2.2.2
    # via requests
werkzeug==3.0.6
    # via flask
wrapt==1.15.0
    # via astroid
python-dotenv==1.0.1
